Change warehouseDB to warehouseDBLegacy
